TST: dtype issues on windows with test_getitem_dups

+++ pandas/core/series.py
@@ -484,7 +484,9 @@ class Series(base.IndexOpsMixin, generic.NDFrame):
                     # we need to box if we have a non-unique index here
                     # otherwise have inline ndarray/lists
                     if not self.index.is_unique:
-                        result = self._constructor(result,index=[key]*len(result)).__finalize__(self)
+                        result = self._constructor(result,
+                                                   index=[key]*len(result)
+                                                   ,dtype=self.dtype).__finalize__(self)
 
             return result
         except InvalidIndexError:
